SDK-44
1. Created a new library repo without previous library versions
2. Changed amount to double
3. Removed a status code
4. Localization of the phone number. Preselects country code based on country parameter passed
4. Buildconfig version added

+++ sdk/src/main/java/com/pesapal/sdk/fragment/mobile_money/mpesa/success/MpesaSuccessFragment.kt
@@ -144,7 +144,7 @@ class MpesaSuccessFragment : Fragment() {
         binding.iconPayMethod.setImageResource(paymentIcon)
 
         binding.tvCurrency.text = transactionStatusResponse.currency
-        binding.tvAmount.text = GeneralUtil.formatAmountText(transactionStatusResponse.amount.toDouble())
+        binding.tvAmount.text = GeneralUtil.formatAmountText(transactionStatusResponse.amount)
         binding.tvConfirmationCode.text = transactionStatusResponse.confirmationCode
         binding.tvMerchantRef.text = transactionStatusResponse.merchantReference
         binding.tvNumberOrCard.text = transactionStatusResponse.paymentAccount
